stock_ai_agent/requirements.txt
groq
pandas
numpy
matplotlib
requests
beautifulsoup4
yfinance
faiss-cpu
sentence-transformers
sqlite-utils
python-dotenv
